# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lumière Beauty API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get user cart
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemBody"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Cart cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /cart/{itemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ORDERS
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get user orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ADMIN ORDERS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [customer, admin]
        skinType:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, email, firstName, lastName, role, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        accessToken:
          type: string
      required: [user, accessToken]
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        skinType:
          type: string
          nullable: true
      required: [email, password, firstName, lastName]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        subcategory:
          type: string
        description:
          type: string
        price:
          type: number
        salePrice:
          type: number
          nullable: true
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        stock:
          type: integer
        skinTypes:
          type: array
          items:
            type: string
        concerns:
          type: array
          items:
            type: string
        isBestSeller:
          type: boolean
        isNewArrival:
          type: boolean
        isFeatured:
          type: boolean
        ingredients:
          type: string
          nullable: true
        howToUse:
          type: string
          nullable: true
      required: [id, name, slug, category, subcategory, description, price, images, rating, reviewCount, stock, skinTypes, concerns, isBestSeller, isNewArrival, isFeatured]
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
      required: [id, productId, quantity, product]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        shipping:
          type: number
        total:
          type: number
        itemCount:
          type: integer
      required: [items, subtotal, shipping, total, itemCount]
    CartItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required: [quantity]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        price:
          type: number
        quantity:
          type: integer
      required: [id, productId, productName, productImage, price, quantity]
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shipping:
          type: number
        total:
          type: number
        shippingAddress:
          $ref: "#/components/schemas/Address"
        createdAt:
          type: string
        userId:
          type: integer
        userEmail:
          type: string
          nullable: true
      required: [id, orderNumber, status, items, subtotal, shipping, total, shippingAddress, createdAt, userId]
    CreateOrderBody:
      type: object
      properties:
        shippingAddress:
          $ref: "#/components/schemas/Address"
        discountCode:
          type: string
          nullable: true
      required: [shippingAddress]
...
    Address:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string
      required: [fullName, street, city, state, zip, country]
...